applications/facenet/src/requirements.txt
tensorflow==1.13.1
asyncio-nats-client
protobuf
